requirements.txt
fastapi
uvicorn
openai
langchain
faiss-cpu
pydantic
python-multipart
sentence-transformers
transformers
python-docx
python-pptx
pypdf
